coredata: move MutableKeyedOptionDict to options
