+++ mesonbuild/compilers/swift.py
@@ -14,7 +14,7 @@ from .compilers import Compiler, clike_debug_args
 
 if T.TYPE_CHECKING:
     from .. import build
-    from ..coredata import MutableKeyedOptionDictType
+    from ..options import MutableKeyedOptionDictType
     from ..dependencies import Dependency
     from ..envconfig import MachineInfo
     from ..environment import Environment
